+++ sdk/dotnet/Grant.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi.Serialization;
 
-namespace Pulumi.Postgresql
+namespace Pulumi.PostgreSql
 {
     /// <summary>
     /// The ``postgresql..Grant`` resource creates and manages privileges given to a user for a database schema.