+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD (Unreleased)
-_(none)_
+* Namespace names in .NET SDK are adjusted to PascalCase
+([#28](https://github.com/pulumi/pulumi-postgresql/pull/28)).
 
 ---