+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## HEAD (Unreleased)
+_(none)_
+
+---
+
+## 1.3.0
 * Namespace names in .NET SDK are adjusted to PascalCase
 ([#28](https://github.com/pulumi/pulumi-postgresql/pull/28)).
 * Upgrade to pulumi-terraform-bridge v1.5.2
 * Upgrade to v1.4.0 of the Postgresql Terraform Provider
-
----
 
 ## 1.2.0 (2019-12-04)
 * Upgrade to support go 1.13.x